@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<IncludeBuildOutput>true</IncludeBuildOutput>
 		<ProduceReferenceAssembly>false</ProduceReferenceAssembly>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
@@ -18,6 +18,7 @@
 		<PackageLicenseExpression>MIT</PackageLicenseExpression>
 		<PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
 		<PackageReleaseNotes>
+			* 2023-11-05 - 0.1.4-Alpha: Introduce BindingTypeConverterExtender, more bug fixes.
 			* 2023-11-05 - 0.1.3-Alpha: Bug fixes in SymbolImageFactory and SymbolToolStripRenderer.
 			* 2023-11-05 - 0.1.2-Alpha: Added rudimentary dark-mode handling for ToolStrips.
 			* 2023-11-04 - 0.1.1-Alpha: Added SymbolImageFactory and look-up enums for both MDL2 and Segoe Fluent Symbols.
