@@ -1,7 +1,11 @@
 ﻿namespace WinForms.PowerTools.Components;
 
+/// <summary>
+///  Defines the theming mode.
+/// </summary>
 public enum ThemingMode
 {
     LightMode,
-    DarkMode
+    DarkMode,
+    System
 }
