@@ -28,32 +28,94 @@
         /// </summary>
         private void InitializeComponent()
         {
+            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(ControlsTestForm));
             menuStrip1 = new System.Windows.Forms.MenuStrip();
-            toolStripFluentSymbolMenuItem2 = new PowerTools.Controls.ToolStripFluentSymbolMenuItem();
-            menuStrip1.SuspendLayout();
+            statusStrip1 = new System.Windows.Forms.StatusStrip();
+            toolStripSplitButton1 = new System.Windows.Forms.ToolStripSplitButton();
+            groupBox1 = new System.Windows.Forms.GroupBox();
+            radioButton3 = new System.Windows.Forms.RadioButton();
+            radioButton2 = new System.Windows.Forms.RadioButton();
+            radioButton1 = new System.Windows.Forms.RadioButton();
+            statusStrip1.SuspendLayout();
+            groupBox1.SuspendLayout();
             SuspendLayout();
             // 
             // menuStrip1
             // 
+            menuStrip1.BackColor = System.Drawing.SystemColors.ButtonShadow;
+            menuStrip1.Dock = System.Windows.Forms.DockStyle.Left;
             menuStrip1.ImageScalingSize = new System.Drawing.Size(20, 20);
-            menuStrip1.Items.AddRange(new System.Windows.Forms.ToolStripItem[] { toolStripFluentSymbolMenuItem2 });
             menuStrip1.Location = new System.Drawing.Point(0, 0);
             menuStrip1.Name = "menuStrip1";
-            menuStrip1.Size = new System.Drawing.Size(800, 92);
+            menuStrip1.RenderMode = System.Windows.Forms.ToolStripRenderMode.Professional;
+            menuStrip1.Size = new System.Drawing.Size(30, 450);
             menuStrip1.TabIndex = 0;
             menuStrip1.Text = "menuStrip1";
             // 
-            // toolStripFluentSymbolMenuItem2
+            // statusStrip1
             // 
-            toolStripFluentSymbolMenuItem2.ImageScaling = System.Windows.Forms.ToolStripItemImageScaling.None;
-            toolStripFluentSymbolMenuItem2.Name = "toolStripFluentSymbolMenuItem2";
-            toolStripFluentSymbolMenuItem2.ScalePercentage = 100;
-            toolStripFluentSymbolMenuItem2.Size = new System.Drawing.Size(78, 88);
-            toolStripFluentSymbolMenuItem2.Symbol = PowerTools.Controls.SegoeFluentIcons.Calculator;
-            toolStripFluentSymbolMenuItem2.SymbolColor = System.Drawing.Color.FromArgb(0, 0, 192);
-            toolStripFluentSymbolMenuItem2.SymbolSize = new System.Drawing.Size(64, 64);
-            toolStripFluentSymbolMenuItem2.Text = "Test";
-            toolStripFluentSymbolMenuItem2.TextImageRelation = System.Windows.Forms.TextImageRelation.ImageAboveText;
+            statusStrip1.BackColor = System.Drawing.SystemColors.ButtonShadow;
+            statusStrip1.ImageScalingSize = new System.Drawing.Size(20, 20);
+            statusStrip1.Items.AddRange(new System.Windows.Forms.ToolStripItem[] { toolStripSplitButton1 });
+            statusStrip1.Location = new System.Drawing.Point(30, 424);
+            statusStrip1.Name = "statusStrip1";
+            statusStrip1.Size = new System.Drawing.Size(770, 26);
+            statusStrip1.TabIndex = 1;
+            statusStrip1.Text = "statusStrip1";
+            // 
+            // toolStripSplitButton1
+            // 
+            toolStripSplitButton1.DisplayStyle = System.Windows.Forms.ToolStripItemDisplayStyle.Image;
+            toolStripSplitButton1.Image = (System.Drawing.Image)resources.GetObject("toolStripSplitButton1.Image");
+            toolStripSplitButton1.ImageTransparentColor = System.Drawing.Color.Magenta;
+            toolStripSplitButton1.Name = "toolStripSplitButton1";
+            toolStripSplitButton1.Size = new System.Drawing.Size(39, 24);
+            toolStripSplitButton1.Text = "toolStripSplitButton1";
+            // 
+            // groupBox1
+            // 
+            groupBox1.Controls.Add(radioButton3);
+            groupBox1.Controls.Add(radioButton2);
+            groupBox1.Controls.Add(radioButton1);
+            groupBox1.Location = new System.Drawing.Point(241, 131);
+            groupBox1.Name = "groupBox1";
+            groupBox1.Size = new System.Drawing.Size(320, 144);
+            groupBox1.TabIndex = 2;
+            groupBox1.TabStop = false;
+            groupBox1.Text = "Theming";
+            // 
+            // radioButton3
+            // 
+            radioButton3.AutoSize = true;
+            radioButton3.Location = new System.Drawing.Point(92, 98);
+            radioButton3.Name = "radioButton3";
+            radioButton3.Size = new System.Drawing.Size(77, 24);
+            radioButton3.TabIndex = 2;
+            radioButton3.TabStop = true;
+            radioButton3.Text = "System";
+            radioButton3.UseVisualStyleBackColor = true;
+            // 
+            // radioButton2
+            // 
+            radioButton2.AutoSize = true;
+            radioButton2.Location = new System.Drawing.Point(92, 68);
+            radioButton2.Name = "radioButton2";
+            radioButton2.Size = new System.Drawing.Size(104, 24);
+            radioButton2.TabIndex = 1;
+            radioButton2.TabStop = true;
+            radioButton2.Text = "Dark Mode";
+            radioButton2.UseVisualStyleBackColor = true;
+            // 
+            // radioButton1
+            // 
+            radioButton1.AutoSize = true;
+            radioButton1.Location = new System.Drawing.Point(92, 38);
+            radioButton1.Name = "radioButton1";
+            radioButton1.Size = new System.Drawing.Size(106, 24);
+            radioButton1.TabIndex = 0;
+            radioButton1.TabStop = true;
+            radioButton1.Text = "Light Mode";
+            radioButton1.UseVisualStyleBackColor = true;
             // 
             // ControlsTestForm
             // 
@@ -61,12 +123,16 @@
             AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             AutoSize = true;
             ClientSize = new System.Drawing.Size(800, 450);
+            Controls.Add(groupBox1);
+            Controls.Add(statusStrip1);
             Controls.Add(menuStrip1);
             MainMenuStrip = menuStrip1;
             Name = "ControlsTestForm";
-            Text = "DarkModeTestForm2";
-            menuStrip1.ResumeLayout(false);
-            menuStrip1.PerformLayout();
+            Text = "1st Controls Test Form";
+            statusStrip1.ResumeLayout(false);
+            statusStrip1.PerformLayout();
+            groupBox1.ResumeLayout(false);
+            groupBox1.PerformLayout();
             ResumeLayout(false);
             PerformLayout();
         }
@@ -74,9 +140,14 @@
         #endregion
 
         private PowerTools.Components.ThemingComponent themingComponent1;
-        private PowerTools.Controls.ToolStripFluentSymbolMenuItem toolStripFluentSymbolMenuItem1;
+        private PowerTools.Controls.ToolStripSymbolMenuItem toolStripFluentSymbolMenuItem1;
         private PowerTools.Components.ThemingComponent themingComponent2;
         private System.Windows.Forms.MenuStrip menuStrip1;
-        private PowerTools.Controls.ToolStripFluentSymbolMenuItem toolStripFluentSymbolMenuItem2;
+        private System.Windows.Forms.StatusStrip statusStrip1;
+        private System.Windows.Forms.ToolStripSplitButton toolStripSplitButton1;
+        private System.Windows.Forms.GroupBox groupBox1;
+        private System.Windows.Forms.RadioButton radioButton3;
+        private System.Windows.Forms.RadioButton radioButton2;
+        private System.Windows.Forms.RadioButton radioButton1;
     }
 }
